perf: PerfIfaceShell — the duck-typing shells, executed under Native AOT
The runtime duck-typing path is the one operation the converted runtime answers with
neither a nominal adapter nor generated code, and until now nothing in the repo had ever
EXECUTED it in a Native AOT binary. That mattered because it is precisely the path whose
AOT behavior cannot be reasoned about from source: ilc roots the instantiations it can see,
and every tier decision in AdapterBinder turns on whether a run-time-driven
MakeGenericType is one of them.

PerfIfaceShell asserts twice per iteration against an interface literal — once on a
value-typed dynamic value (a named int with a value-receiver method), once on a
pointer-sourced one (a struct with a pointer-receiver method) — so one iteration exercises
BOTH tiers. An anonymous interface is used because it reaches the shell tier by
construction in a single package: neither compile-time recorder covers an interface literal,
so no adapter can exist for either pair. The mechanism reached is the same one a named
interface uses — same emitted shells, same binder, same tier choice.

What the AOT column establishes (measured on this machine, IsDynamicCodeSupported=False):

  * value-typed  -> Δrun_typeᴛObj      (reflective object shell, both builds)
  * pointer-sourced -> Δrun_typeᴛ1<box> on the JIT, Δrun_typeᴛ1ᴛObj under AOT

so the belt is OBSERVED FIRING, and firing exactly where the design said it might: the
generic shell closes over the POINTEE, a struct here, whose instantiation ilc has not
rooted — and the binder degrades it to the object shell rather than to a miss. The checksum
matches Go in all three variants, which the Verify phase requires before anything is timed.

An A/B against the pre-Stage-3 mechanism on the same program is what makes this a gate and
not a curiosity: with the ᴛAs/MakeGenericMethod route, the AOT binary answers MISS for BOTH
tiers and prints a checksum of 0 — silently computing the wrong answer — and takes 9.8 s
doing it, because a failing close is retried every iteration. 202 ms and a correct checksum
now.

Results block regenerated for all ten benchmarks so the table is internally consistent
(same converter, same session); the prior 2026-07-12 / SDK 9.0.315 table moves to History
per the README's own convention. Two rows moved for unrelated reasons already on master:
Channel 3.39x -> 1.87x (the channels redesign) and Map/Startup within noise. The new row is
~193x on the JIT and ~72x under AOT, and the reading notes say plainly why that is the price
of a capability rather than a regression — Go answers this with a ~1.3 ns cached itab
lookup, a recorded pair resolves nominally at ~1.1 ns and never comes near this path, and a
shell obtained once costs 4.4 ns/call on the delegate tier.

Also here, all mechanical: the benchmark projects' generated .cs/.csproj are re-baselined
to current converter output (they were last regenerated 2026-07-12 — the drift is the
channels cap-0 rendering, untyped-const casting at the reference, and the shared csproj
template), and PerfIfaceShell plus the never-registered PerfStringView are added to
go2cs.slnx (the harness builds by path, so an unregistered project passes the suite but
loses its Debug/$(go2csPath) context in Visual Studio).

Gate: full performance suite PASS — 10/10 transpiled, built three ways, output verified
identical across Go / C# JIT / C# Native AOT, measured (164 s).

(cherry picked from commit fd0c2f95952affe513914dfa706033d08a07276a)

+++ src/Tests/Performance/PerfStartup/PerfStartup.cs
@@ -5,8 +5,8 @@ using fmt = fmt_package;
 partial class main_package {
 
 internal static void Main() {
-    fmt.Println("checksum:", 42);
-    fmt.Println("elapsed_ns:", 0);
+    fmt.Println("checksum:", (nint)(42));
+    fmt.Println("elapsed_ns:", (nint)(0));
 }
 
 } // end main_package
